Add LoadingSpinner to improve loading states
Introduced a new reusable LoadingSpinner component and replaced all previous loading state indications with it. The new component provides a better visual cue to users during loading states. Minor changes were also made to the Story arguments for the Button component, and an addition of 'darkRed' to colors in lightTheme.