Update auth provider and adapt code to use isAdmin flag
The commit includes a significant refactor of the auth provider to include the isAdmin flag. This flag, received from the login API, has been stored in local storage along with the token. This new flag has been used to refactor the route determination logic and enable correct navigation from the LandingPage for admin users. Minor changes to other files include adjustment in number of rows in AdminEventsTable, update to Navbar rendering logic based on isAdmin flag and correction in route path navigation in userPanel.