Add image management functionality to events
This commit introduces the ability to add, preview, and delete images related to events. The main changes are the creation of `ImageContainer`, `ImagePreview`, and `ImageComponent` tsx files which handle the display and manipulation of image data. Functions to download and delete images from events have also been added to the events API. Lastly, the event details card has been updated to display image content.